Fix bugs, refactor code and add preferences

+++ app/src/main/java/info/nightscout/androidaps/db/CareportalEvent.java
@@ -86,8 +86,8 @@ public class CareportalEvent implements DataPointWithLabelInterface {
         return System.currentTimeMillis() - date;
     }
 
-    public long getHoursFromStart() {
-        return (System.currentTimeMillis() - date) / (60 * 60 * 1000);
+    public double getHoursFromStart() {
+        return (System.currentTimeMillis() - date) / (60 * 60 * 1000.0);
     }
 
     public String age() {
